chore: split cmd history into rust crate
