+++ Cargo.toml
@@ -30,7 +30,8 @@ rust_debugger = { path = "rust_debugger" }
 rust_pty = { path = "rust_pty" }
 rust_wayland = { path = "rust_wayland" }
 rust_change = { path = "rust_change" }
-rust_excmd = { path = "rust_excmd" }
+rust_excmds = { path = "rust_excmds" }
+rust_cmdhist = { path = "rust_cmdhist" }
 rust_alloc = { path = "rust_alloc" }
 rust_message = { path = "rust_message" }
